engine: Fix bug with Risor modules (dependencies)
The modules and builtins must be loaded dynamically - the cannot be defined as
a global variable because Risor modules use dynamic attributes function which
only evaluates once per execution. This was the case for stdin/stdout fields.

Also, OS context must be propagated to compilation and run methods, therefore
OS initialization is again left to a caller, reverting the previous changes.