commands/job/create: Do not copy files to agent's storage
The job is created using the arguments provided by a user on the command line.
The contents of the arguments is treated as an opaque string that is split into
parts. The first part is treated as a command, the rest are its arguments.

Previously the first part was treated as a local file which was copied to
agent's storage. This approach only works with vessel and would not work with
custom agents.